Update README.md
For this Assignment you are going to find and run the Help Commends on 20 unique Cmdlets. It is recommended that you watch the lecture on the Help System to see how to use the Help Command and Get Command Cmdlets in your Windows Powershell runtime environment.

In your Windows VM:

Start Powershell and run the cmdlet [Start-transcript]
Run the Help Command on your chosen Cmdlet
Explore at least 2 different areas of that Cmdlet
Repeat for each of your 20 different Cmdlets
Run the cmdlet [Stop-transcript]
Go to file location where the .txt is stored
Rename the file to "FirstName-LastName-Assignment6a.txt"
Upload file to this Assignment section before the Deadline
I have linked to Microsoft documentation on how Start-Transcript and Stop-Transcript work below:

Start-Transcript
Link: https://docs.microsoft.com/en-us/powershell/module/microsoft.powershell.host/start-transcript?view=powershell-7 (Links to an external site.)

Stop-Transcript
Link: https://docs.microsoft.com/en-us/powershell/module/microsoft.powershell.host/stop-transcript?view=powershell-7 (Links to an external site.)

Note: Unless you specify where the file is stored, [Start-Transcript] stores the transcripts in the $Home\My Documents directory as \PowerShell_transcript.<time-stamp>.txt files.